requirements.txt
asgiref==3.2.3
cachetools==4.2.0
certifi==2019.11.28
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
defusedxml==0.6.0
Django==2.2.6
django-appconf==1.0.4
django-ckeditor==6.0.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-simple-mail==2.3.3
html2text==2020.1.16
idna==2.8
lxml==4.6.2
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
pilkit==2.0
Pillow==8.0.1
pipfile==0.0.2
premailer==3.7.0
psycopg2==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
SQLAlchemy==1.3.15
sqlparse==0.3.0
toml==0.10.2
urllib3==1.25.7
